Server/requirements.txt
Flask==2.3.2
numpy==1.23.5
tensorflow==2.12.0
python-dotenv==1.0.0
